requirements.txt
playwright
pandas
python-dotenv
flask
